seperate histogram tests, move init Re #151
